[Fix] title 삭제
